+++ middleware_api/lambda/checkpoints/list_checkpoints.py
@@ -2,12 +2,10 @@ import json
 import logging
 import os
 
-from multi_users._types import PARTITION_KEYS, Role
-
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok
-from lib._types import CheckPoint
-from multi_users.utils import get_user_roles, check_user_permissions, get_permissions_by_username
+from common.types import CheckPoint, PARTITION_KEYS, Role
+from common.utils import get_user_roles, check_user_permissions, get_permissions_by_username
 
 checkpoint_table = os.environ.get('CHECKPOINT_TABLE')
 bucket_name = os.environ.get('S3_BUCKET')
