+++ middleware_api/lambda/datasets/create_dataset.py
@@ -5,11 +5,11 @@ from dataclasses import dataclass
 from datetime import datetime
 from typing import Any, List
 
-from _types import DatasetItem, DatasetInfo, DatasetStatus, DataStatus
+from common.types import DatasetItem, DatasetInfo, DatasetStatus, DataStatus
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request, internal_server_error
 from common.util import get_s3_presign_urls
-from multi_users.utils import get_permissions_by_username, get_user_roles
+from common.utils import get_permissions_by_username, get_user_roles
 
 dataset_item_table = os.environ.get('DATASET_ITEM_TABLE')
 dataset_info_table = os.environ.get('DATASET_INFO_TABLE')
