+++ middleware_api/lambda/datasets/get_dataset.py
@@ -1,11 +1,11 @@
 import logging
 import os
 
-from _types import DatasetItem, DatasetInfo
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, not_found, forbidden
 from common.util import generate_presign_url
-from multi_users.utils import get_permissions_by_username, get_user_roles, check_user_permissions
+from common.types import DatasetItem, DatasetInfo
+from common.utils import get_permissions_by_username, get_user_roles, check_user_permissions
 
 dataset_item_table = os.environ.get('DATASET_ITEM_TABLE')
 dataset_info_table = os.environ.get('DATASET_INFO_TABLE')
