+++ middleware_api/lambda/endpoints/create_endpoint.py
@@ -7,11 +7,11 @@ from datetime import datetime
 
 import boto3
 
-from lib._enums import EndpointStatus
-from _types import EndpointDeploymentJob
 from common.ddb_service.client import DynamoDbUtilsService
+from common.enums import EndpointStatus
 from common.response import ok, bad_request
-from multi_users.utils import get_permissions_by_username
+from common.types import EndpointDeploymentJob
+from common.utils import get_permissions_by_username
 
 sagemaker_endpoint_table = os.environ.get('DDB_ENDPOINT_DEPLOYMENT_TABLE_NAME')
 user_table = os.environ.get('MULTI_USER_TABLE')
