+++ middleware_api/lambda/inference_v2/inference_api.py
@@ -11,12 +11,12 @@ from sagemaker.deserializers import JSONDeserializer
 from sagemaker.predictor_async import AsyncPredictor
 from sagemaker.serializers import JSONSerializer
 
-from _types import InferenceJob, InvocationsRequest, EndpointDeploymentJob
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request
 from common.util import generate_presign_url, load_json_from_s3, upload_json_to_s3, split_s3_path
-from lib._types import CheckPoint, CheckPointStatus
-from multi_users.utils import get_user_roles, check_user_permissions
+from common.types import CheckPoint, CheckPointStatus
+from common.types import InferenceJob, InvocationsRequest, EndpointDeploymentJob
+from common.utils import get_user_roles, check_user_permissions
 
 bucket_name = os.environ.get('S3_BUCKET')
 checkpoint_table = os.environ.get('CHECKPOINT_TABLE')
