+++ middleware_api/lambda/models/create_model.py
@@ -11,10 +11,10 @@ from botocore.exceptions import ClientError
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request, internal_server_error
 from common.stepfunction_service.client import StepFunctionUtilsService
-from lib._types import Model, CreateModelStatus, CheckPoint, CheckPointStatus, MultipartFileReq
-from lib.common_tools import get_base_model_s3_key, get_base_checkpoint_s3_key, \
+from common.types import Model, CreateModelStatus, CheckPoint, CheckPointStatus, MultipartFileReq
+from common.common_tools import get_base_model_s3_key, get_base_checkpoint_s3_key, \
     batch_get_s3_multipart_signed_urls
-from multi_users.utils import get_permissions_by_username, get_user_roles
+from common.utils import get_permissions_by_username, get_user_roles
 
 bucket_name = os.environ.get('S3_BUCKET')
 model_table = os.environ.get('DYNAMODB_TABLE')
