+++ middleware_api/lambda/models/list_models.py
@@ -4,8 +4,8 @@ import os
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request
 from common.stepfunction_service.client import StepFunctionUtilsService
-from lib._types import Model
-from multi_users.utils import get_permissions_by_username, get_user_roles, check_user_permissions
+from common.types import Model
+from common.utils import get_permissions_by_username, get_user_roles, check_user_permissions
 
 bucket_name = os.environ.get('S3_BUCKET')
 model_table = os.environ.get('DYNAMODB_TABLE')
