+++ middleware_api/lambda/roles/list_roles.py
@@ -4,8 +4,8 @@ import os
 
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok
-from lib._types import Role, PARTITION_KEYS
-from utils import get_permissions_by_username, get_user_roles
+from common.types import Role, PARTITION_KEYS
+from common.utils import get_permissions_by_username, get_user_roles
 
 user_table = os.environ.get('MULTI_USER_TABLE')
