+++ middleware_api/lambda/trainings/create_training_job.py
@@ -10,16 +10,17 @@ import boto3
 
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request, not_found, forbidden, internal_server_error
+from common.types import TrainJob, TrainJobStatus, Model, CreateModelStatus, CheckPoint, CheckPointStatus
 from common.util import get_s3_presign_urls
 from common.util import load_json_from_s3, save_json_to_file
-from lib._types import TrainJob, TrainJobStatus, Model, CreateModelStatus, CheckPoint, CheckPointStatus
-from multi_users.utils import get_permissions_by_username, get_user_roles
+from common.utils import get_permissions_by_username, get_user_roles
 
 bucket_name = os.environ.get('S3_BUCKET')
 train_table = os.environ.get('TRAIN_TABLE')
 model_table = os.environ.get('MODEL_TABLE')
 checkpoint_table = os.environ.get('CHECKPOINT_TABLE')
 user_table = os.environ.get('MULTI_USER_TABLE')
+
 logger = logging.getLogger('boto3')
 logger.setLevel(logging.INFO)
 ddb_service = DynamoDbUtilsService(logger=logger)
