+++ middleware_api/lambda/users/delete_users.py
@@ -4,9 +4,9 @@ import os
 
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok
-from lib._types import PARTITION_KEYS
+from common.types import PARTITION_KEYS
 from users.create_user import _check_action_permission
-from utils import KeyEncryptService
+from common.utils import KeyEncryptService
 
 user_table = os.environ.get('MULTI_USER_TABLE')
 kms_key_id = os.environ.get('KEY_ID')
